Basic implementation of Yomi library; Skullgirls/Filia demo
yomi-lib.rkt is a step above the uinput code. It introduces
concepts like moves, combos, and delays. It provides functions for
executing combos/moves on the virutal input device and describing
links/cancels.

skullgirls-filia-demo.rkt is a demonstration of how to describe a
game and character at the library level, as well as how to write
combos.