Implement mock device; restructure project (#3)
The mock device simply keeps track of what commands get sent to it.
Nothing actually gets performed, making it suitable for development
on unsupported platforms. The command list can be accessed through
received-inputs.

This PR also restructures the project so the code is in a lib folder,
and unit tests are in a test folder. There are also updates to the
Makefile to help with testing and development.