Tests for yomi-lib; refactor delays (#4)
Add tests for the yomi-lib module. Some changes needed to be made
to the Delay data definition in order to make testing possible.
Instead of thunks, Delays are now 1-field structures, and the call to
'sleep' is made in perform-combo. This also makes the code more
readable.

The Filia demo still works with these changes.