Use angle brackets instead of parens for delays
To use parens in identifiers, we'd have to modify the reader. We've
looked into this a bit and it's too involved for such a small piece of
behavior. Luckily, the reader does not treat < and > like other
brackets, so we can just use those and they work in identifiers just
fine.

This commit includes code changes and updates to the design docs. We
originally wanted to use > for cancels; I've updated it to use ~ but
we should have a conversation about this later.