Tests for link/cancel delays in combos (#8)
(Fair warning, this is kind of hacky.)

Add perform-combo-for-test. The only difference between it and
perform-combo is that it writes delays to the mock device's received
inputs instead of sleeping. This allows us to test links/cancels in
combos, but it can't help us with delays for held inputs moving
forward.